requirements.txt
Django>=6.0
djangorestframework>=3.14
drf-spectacular>=0.27
django-cors-headers>=4.3
Pillow>=10.0
Pillow>=10.0
